+++ packages/core/src/node/runtimeModule/routeData.ts
@@ -1,5 +1,4 @@
 import { addLeadingSlash } from '@rspress/shared';
-import { RspackVirtualModulePlugin } from 'rspack-plugin-virtual-module';
 import { FactoryContext, RuntimeModuleID } from '.';
 
 export const normalizeRoutePath = (routePath: string) => {
@@ -10,9 +9,9 @@ export const normalizeRoutePath = (routePath: string) => {
 export async function routeVMPlugin(context: FactoryContext) {
   const { routeService } = context;
   // client: The components of route is lazy loaded
-  const plugin = new RspackVirtualModulePlugin({
+
+  return {
     [RuntimeModuleID.RouteForClient]: routeService.generateRoutesCode(false),
     [RuntimeModuleID.RouteForSSR]: routeService.generateRoutesCode(true),
-  });
-  return plugin;
+  };
 }