+++ packages/core/src/theme-default/layout/Layout/index.tsx
@@ -48,8 +48,10 @@ export const Layout: React.FC<LayoutProps> = props => {
   } = page;
   const localesData = useLocaleSiteData();
   const defaultLang = siteData.lang || '';
+  // Priority: frontmatter.navbar > themeConfig.hideNavbar
+  // Always show sidebar by default
   const hideNavbar =
-    frontmatter?.hideNavbar ?? themeConfig?.hideNavbar ?? false;
+    !(frontmatter?.navbar ?? true) || (themeConfig?.hideNavbar ?? false);
   // Priority: front matter title > h1 title
   let title = (frontmatter?.title as string) ?? articleTitle;
   const mainTitle = siteData.title || localesData.title;
@@ -140,11 +142,22 @@ export const Layout: React.FC<LayoutProps> = props => {
       {bottom}
       {
         // Global UI
-        globalComponents.map((Component, index) => (
-          // The component order is stable
-          // eslint-disable-next-line react/no-array-index-key
-          <Component key={index} />
-        ))
+        globalComponents.map((componentInfo, index) => {
+          if (Array.isArray(componentInfo)) {
+            const [component, props] = componentInfo;
+            return React.createElement(component, {
+              // The component order is stable
+              // eslint-disable-next-line react/no-array-index-key
+              key: index,
+              ...props,
+            });
+          } else {
+            return React.createElement(componentInfo, {
+              // eslint-disable-next-line react/no-array-index-key
+              key: index,
+            });
+          }
+        })
       }
     </div>
   );