+++ packages/create-rspress/src/index.ts
@@ -3,11 +3,35 @@ import path from 'path';
 import { cac } from 'cac';
 import prompts from 'prompts';
 import fs from 'fs-extra';
-import { copyFolder, formatTargetDir, getPkgManager } from './utils';
+import {
+  copyFolder,
+  initSitePrompts,
+  formatTargetDir,
+  getPkgManager,
+  renderFile,
+} from './utils';
+import { CustomPromptObject } from './types';
 
 const cli = cac('create-rspress').help();
 
 cli.command('', 'Create a new rspress site').action(async () => {
+  const siteOptions: CustomPromptObject[] = [
+    {
+      name: 'siteTitle',
+      type: 'text',
+      message: 'Site Title',
+      initial: 'Rspress',
+      value: '',
+    },
+    {
+      name: 'siteDesc',
+      type: 'text',
+      message: 'Site Description',
+      initial: 'Rspack-based Static Site Generator',
+      value: '',
+    },
+  ];
+
   const defaultProjectName = 'rspress-site';
   let targetDir = defaultProjectName;
   const promptProjectDir = async () =>
@@ -33,9 +57,21 @@ cli.command('', 'Create a new rspress site').action(async () => {
     root = path.resolve(process.cwd(), targetDir);
   }
 
+  await initSitePrompts(siteOptions);
+
   await fs.mkdir(root, { recursive: true });
   const srcFolder = path.resolve(__dirname, '../template');
   await copyFolder(srcFolder, targetDir);
+
+  const filesToInterpret = ['docs/index.md', 'rspress.config.ts'];
+  const siteData = siteOptions.reduce((prev: Record<string, any>, cur) => {
+    prev[cur.name as string] = cur.value;
+    return prev;
+  }, {});
+  for (const file of filesToInterpret) {
+    await renderFile(file, targetDir, targetDir, siteData);
+  }
+
   const pkgManager = getPkgManager();
   console.log('\nDone. Now run:\n');
   console.log(`cd ${targetDir}\n`);