+++ packages/create-rspress/src/types.ts
@@ -1,0 +1,6 @@
+import type { PromptObject } from 'prompts';
+
+export interface CustomPromptObject extends PromptObject {
+  value?: any;
+  formatFn?: (...data: any) => any;
+}